Move CUDA platform util to `platforms/cuda.py`
